backend/requirements.txt
tensorflow
Pillow
flask
flask-cors
numpy